requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
comm==0.1.2
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
entrypoints==0.4
executing==1.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
matplotlib-inline==0.1.6
msgpack==1.0.4
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pyzmq==24.0.1
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.8.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
zope.interface==5.5.2
